bug fix. New component Current time
