+++ src/components/CurrentTime.jsx
@@ -1,0 +1,15 @@
+import { useState, useEffect } from "react";
+
+export default function CorrentTime() {
+  const [now, setNow] = useState(new Date());
+
+  useEffect(() => {
+    const interval = setInterval(() => setNow(new Date()), 1000);
+
+    return () => {
+      clearInterval(interval);
+    };
+  }, []);
+
+  return <span>{now.toLocaleTimeString()}</span>;
+}
